refactor(renderer): extract messages panel team mention meta
Move the memoized team mention metadata selector out of MessagesPanel.tsx into a sibling module. The block was already self-contained: two empty-value constants, two local interfaces, three module-level cache variables and three private helpers, all reachable only from the single selector call in the component's store subscription.

The moved code is byte-identical to its previous form; the only change is that the selector now carries `export`. Everything else stays module-private, so visibility narrows rather than widens: `TeamMentionMeta` remains un-exported and is used only as the selector's return type, which is legal under `noEmit`. The module gets a short file-header JSDoc describing the memoization contract.

MessagesPanel.tsx keeps `EMPTY_REPLY_CANDIDATE_MESSAGES` (used elsewhere in the file), drops `TeamSummary` from its `@shared/types` type import, and gains one import of the selector. Prettier then collapses the shortened type import onto a single line. The file goes from 1722 to 1623 lines, staying above the 800-line floor and 99 lines under its frozen cap.

No behaviour change: the cache variables keep the same module-instance lifetime, the comparator and signature encoding are untouched, and the near-duplicate helper in src/renderer/hooks/useStableTeamMentionMeta.ts is deliberately left alone. The existing MessagesPanel test suite passes unmodified.
